+++ package.json
@@ -17,7 +17,7 @@
     "@svgr/webpack": "4.3.3",
     "@typescript-eslint/eslint-plugin": "^2.10.0",
     "@typescript-eslint/parser": "^2.10.0",
-    "axios": "^0.21.1",
+    "axios": "^0.21.4",
     "babel-eslint": "10.0.3",
     "babel-jest": "^24.9.0",
     "babel-loader": "8.0.6",
@@ -84,7 +84,6 @@
     "semver": "6.3.0",
     "streamsaver": "^2.0.4",
     "style-loader": "0.23.1",
-    "swr": "^0.5.6",
     "terser-webpack-plugin": "2.3.4",
     "ts-pnp": "1.1.5",
     "url-loader": "2.3.0",
